Implemented Feed paging
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@1861 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-api/src/main/java/org/sonatype/nexus/feeds/FeedRecorder.java
@@ -117,9 +117,9 @@ public interface FeedRecorder
 
     // reading
 
-    List<Map<String, String>> getEvents( Set<String> types, Set<String> subtypes, Integer from, Integer count );
+    List<Map<String, String>> getEvents( Set<String> types, Set<String> subtypes, Long from, Integer count );
 
-    List<NexusArtifactEvent> getNexusArtifectEvents( Set<String> subtypes, Integer from, Integer count );
+    List<NexusArtifactEvent> getNexusArtifectEvents( Set<String> subtypes, Long from, Integer count );
 
-    List<SystemEvent> getSystemEvents( Set<String> subtypes, Integer from, Integer count );
+    List<SystemEvent> getSystemEvents( Set<String> subtypes, Long from, Integer count );
 }
